backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Session==0.8.0
SQLAlchemy==2.0.36
requests==2.31.0
PyJWT==2.8.0
Authlib==1.3.2
python-dotenv==1.0.1
redis==5.0.8
boto3==1.35.21
python-jose==3.3.0
PyMySQL==1.0.2
gunicorn==20.1.0
twilio==9.3.7
flask_restx==1.3.0
flask_cors==4.0.0
notion-client==2.3.0
docusign-esign==4.0.0
xhtml2pdf